102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	signed float a = 1;
-	signed float b = 2;
-	signed float result;
+	long float a = 1;
+	long float b = 2;
+	long float result;
 	int i;
 
 	for (i = 0; i <= 49; i++)
 	{
-		printf("%.0f, ", signed  a);
+		printf("%.0lf, ", signed  a);
 		result = a + b;
 		a = b;
 		b = result;
